+++ flask_restplus/model.py
@@ -2,7 +2,9 @@
 from __future__ import unicode_literals
 
 import copy
+import inspect
 import re
+import warnings
 
 from collections import MutableMapping
 from six import iteritems, itervalues
@@ -38,11 +40,19 @@ class Model(dict, MutableMapping):
             'name': name
         }
         self.name = name
-        self.__parent__ = None
+        self.__parents__ = []
         self.__mask__ = kwargs.pop('mask', None)
         if self.__mask__ and not isinstance(self.__mask__, Mask):
             self.__mask__ = Mask(self.__mask__)
         super(Model, self).__init__(*args, **kwargs)
+
+        def instance_clone(name, *parents):
+            return self.__class__.clone(name, self, *parents)
+        self.clone = instance_clone
+
+        def instance_inherit(name, *parents):
+            return self.__class__.inherit(name, self, *parents)
+        self.inherit = instance_inherit
 
     @cached_property
     def resolved(self):
@@ -53,8 +63,8 @@ class Model(dict, MutableMapping):
         resolved = copy.deepcopy(self)
 
         # Recursively copy parent fields if necessary
-        if self.__parent__:
-            resolved.update(self.__parent__.resolved)
+        for parent in self.__parents__:
+            resolved.update(parent.resolved)
 
         # Handle discriminator
         candidates = [f for f in itervalues(resolved) if getattr(f, 'discriminator', None)]
@@ -72,23 +82,18 @@ class Model(dict, MutableMapping):
         '''
         Return the ancestors tree
         '''
-        return self.__parent__.tree
-
-    @cached_property
-    def tree(self):
-        '''
-        Return the inheritance tree
-        '''
-        tree = [self.name]
-        return self.ancestors + tree if self.__parent__ else tree
+        ancestors = [p.ancestors for p in self.__parents__]
+        return set.union(set([self.name]), *ancestors)
 
     def get_parent(self, name):
         if self.name == name:
             return self
-        elif self.__parent__:
-            return self.__parent__.get_parent(name)
         else:
-            raise ValueError('Parent ' + name + ' not found')
+            for parent in self.__parents__:
+                found = parent.get_parent(name)
+                if found:
+                    return found
+        raise ValueError('Parent ' + name + ' not found')
 
     @cached_property
     def __schema__(self):
@@ -110,12 +115,14 @@ class Model(dict, MutableMapping):
             'x-mask': str(self.__mask__) if self.__mask__ else None,
         })
 
-        if self.__parent__:
+        if self.__parents__:
+            refs = [
+                {'$ref': '#/definitions/{0}'.format(parent.name)}
+                for parent in self.__parents__
+            ]
+
             return {
-                'allOf': [
-                    {'$ref': '#/definitions/{0}'.format(self.__parent__.name)},
-                    schema
-                ]
+                'allOf': refs + [schema]
             }
         else:
             return schema
@@ -126,20 +133,46 @@ class Model(dict, MutableMapping):
 
         :param str name: The new model name
         :param dict fields: The new model extra fields
-        '''
-        model = Model(name, copy.deepcopy(self))
-        model.update(fields)
-        return model
 
-    def inherit(self, name, fields):
+        :depreated: since 0.9. Use :meth:`clone` instead.
+        '''
+        warnings.warn('extend is is deprecated, use clone instead', DeprecationWarning, stacklevel=2)
+        if isinstance(fields, (list, tuple)):
+            return self.clone(name, *fields)
+        else:
+            return self.clone(name, fields)
+
+    @classmethod
+    def clone(cls, name, *parents):
+        '''
+        Clone these models (Duplicate all fields)
+
+        It can be used from the class
+
+        >>> model = Model.clone(fields_1, fields_2)
+
+        or from an Instanciated model
+
+        >>> new_model = model.clone(fields_1, fields_2)
+
+        :param str name: The new model name
+        :param dict parents: The new model extra fields
+        '''
+        fields = {}
+        for parent in parents:
+            fields.update(copy.deepcopy(parent))
+        return cls(name, fields)
+
+    @classmethod
+    def inherit(cls, name, *parents):
         '''
         Inherit this model (use the Swagger composition pattern aka. allOf)
 
         :param str name: The new model name
         :param dict fields: The new model extra fields
         '''
-        model = Model(name, fields)
-        model.__parent__ = self
+        model = Model(name, parents[-1])
+        model.__parents__ = parents[:-1]
         return model
 
     def validate(self, data, resolver=None):
@@ -157,3 +190,8 @@ class Model(dict, MutableMapping):
             path.append(name)
         key = '.'.join(str(p) for p in path)
         return key, error.message
+
+    def __unicode__(self):
+        return 'Model({name},{{{fields}}})'.format(name=self.name, fields=','.join(self.keys()))
+
+    __str__ = __unicode__