+++ tests/test_model.py
@@ -1,21 +1,14 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from flask import Blueprint
-from flask_restplus import fields, Api
+from flask_restplus import fields, Model
 
 from . import TestCase
 
 
 class ModelTestCase(TestCase):
-    def setUp(self):
-        super(ModelTestCase, self).setUp()
-        blueprint = Blueprint('api', __name__)
-        self.api = Api(blueprint)
-        self.app.register_blueprint(blueprint)
-
     def test_model_as_flat_dict(self):
-        model = self.api.model('Person', {
+        model = Model('Person', {
             'name': fields.String,
             'age': fields.Integer,
             'birthdate': fields.DateTime,
@@ -37,11 +30,11 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_as_nested_dict(self):
-        address = self.api.model('Address', {
+        address = Model('Address', {
             'road': fields.String,
         })
 
-        person = self.api.model('Person', {
+        person = Model('Person', {
             'name': fields.String,
             'age': fields.Integer,
             'birthdate': fields.DateTime,
@@ -76,7 +69,7 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_as_dict_with_list(self):
-        model = self.api.model('Person', {
+        model = Model('Person', {
             'name': fields.String,
             'age': fields.Integer,
             'tags': fields.List(fields.String),
@@ -100,11 +93,11 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_as_nested_dict_with_list(self):
-        address = self.api.model('Address', {
+        address = Model('Address', {
             'road': fields.String,
         })
 
-        person = self.api.model('Person', {
+        person = Model('Person', {
             'name': fields.String,
             'age': fields.Integer,
             'birthdate': fields.DateTime,
@@ -142,7 +135,7 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_with_required(self):
-        model = self.api.model('Person', {
+        model = Model('Person', {
             'name': fields.String(required=True),
             'age': fields.Integer,
             'birthdate': fields.DateTime(required=True),
@@ -165,11 +158,11 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_as_nested_dict_and_required(self):
-        address = self.api.model('Address', {
+        address = Model('Address', {
             'road': fields.String,
         })
 
-        person = self.api.model('Person', {
+        person = Model('Person', {
             'name': fields.String,
             'age': fields.Integer,
             'birthdate': fields.DateTime,
@@ -204,7 +197,7 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_with_discriminator(self):
-        model = self.api.model('Person', {
+        model = Model('Person', {
             'name': fields.String(discriminator=True),
             'age': fields.Integer,
         })
@@ -219,7 +212,7 @@ class ModelTestCase(TestCase):
         })
 
     def test_model_with_discriminator_override_require(self):
-        model = self.api.model('Person', {
+        model = Model('Person', {
             'name': fields.String(discriminator=True, required=False),
             'age': fields.Integer,
         })
@@ -233,14 +226,14 @@ class ModelTestCase(TestCase):
             'required': ['name']
         })
 
-    def test_extend(self):
-        parent = self.api.model('Parent', {
+    def test_clone_from_instance(self):
+        parent = Model('Parent', {
             'name': fields.String,
             'age': fields.Integer,
             'birthdate': fields.DateTime,
         })
 
-        child = self.api.extend('Child', parent, {
+        child = parent.clone('Child', {
             'extra': fields.String,
         })
 
@@ -262,21 +255,47 @@ class ModelTestCase(TestCase):
             }
         })
 
-        self.assertIn('Parent', self.api.models)
-        self.assertIn('Child', self.api.models)
-
-    def test_extend_with_multiple_parents(self):
-        grand_parent = self.api.model('GrandParent', {
-            'grand_parent': fields.String,
-        })
-
-        parent = self.api.model('Parent', {
+    def test_clone_from_class(self):
+        parent = Model('Parent', {
             'name': fields.String,
             'age': fields.Integer,
             'birthdate': fields.DateTime,
         })
 
-        child = self.api.extend('Child', [grand_parent, parent], {
+        child = Model.clone('Child', parent, {
+            'extra': fields.String,
+        })
+
+        self.assertEqual(child.__schema__, {
+            'properties': {
+                'name': {
+                    'type': 'string'
+                },
+                'age': {
+                    'type': 'integer'
+                },
+                'birthdate': {
+                    'type': 'string',
+                    'format': 'date-time'
+                },
+                'extra': {
+                    'type': 'string'
+                }
+            }
+        })
+
+    def test_clone_from_instance_with_multiple_parents(self):
+        grand_parent = Model('GrandParent', {
+            'grand_parent': fields.String,
+        })
+
+        parent = Model('Parent', {
+            'name': fields.String,
+            'age': fields.Integer,
+            'birthdate': fields.DateTime,
+        })
+
+        child = grand_parent.clone('Child', parent, {
             'extra': fields.String,
         })
 
@@ -301,17 +320,49 @@ class ModelTestCase(TestCase):
             }
         })
 
-        self.assertIn('GrandParent', self.api.models)
-        self.assertIn('Parent', self.api.models)
-        self.assertIn('Child', self.api.models)
+    def test_clone_from_class_with_multiple_parents(self):
+        grand_parent = Model('GrandParent', {
+            'grand_parent': fields.String,
+        })
 
-    def test_inherit(self):
-        parent = self.api.model('Parent', {
+        parent = Model('Parent', {
+            'name': fields.String,
+            'age': fields.Integer,
+            'birthdate': fields.DateTime,
+        })
+
+        child = Model.clone('Child', grand_parent, parent, {
+            'extra': fields.String,
+        })
+
+        self.assertEqual(child.__schema__, {
+            'properties': {
+                'grand_parent': {
+                    'type': 'string'
+                },
+                'name': {
+                    'type': 'string'
+                },
+                'age': {
+                    'type': 'integer'
+                },
+                'birthdate': {
+                    'type': 'string',
+                    'format': 'date-time'
+                },
+                'extra': {
+                    'type': 'string'
+                }
+            }
+        })
+
+    def test_inherit_from_instance(self):
+        parent = Model('Parent', {
             'name': fields.String,
             'age': fields.Integer,
         })
 
-        child = self.api.inherit('Child', parent, {
+        child = parent.inherit('Child', {
             'extra': fields.String,
         })
 
@@ -322,9 +373,9 @@ class ModelTestCase(TestCase):
             }
         })
         self.assertEqual(child.__schema__, {
-            'allOf': [{
-                    '$ref': '#/definitions/Parent'
-                }, {
+            'allOf': [
+                {'$ref': '#/definitions/Parent'},
+                {
                     'properties': {
                         'extra': {'type': 'string'}
                     }
@@ -332,26 +383,102 @@ class ModelTestCase(TestCase):
             ]
         })
 
-        self.assertIn('Parent', self.api.models)
-        self.assertIn('Child', self.api.models)
-
-    def test_inherit_inline(self):
-        parent = self.api.model('Person', {
+    def test_inherit_from_class(self):
+        parent = Model('Parent', {
             'name': fields.String,
             'age': fields.Integer,
         })
 
-        child = self.api.inherit('Child', parent, {
+        child = Model.inherit('Child', parent, {
             'extra': fields.String,
         })
 
-        self.api.model('Output', {
-            'child': fields.Nested(child),
-            'children': fields.List(fields.Nested(child))
+        self.assertEqual(parent.__schema__, {
+            'properties': {
+                'name': {'type': 'string'},
+                'age': {'type': 'integer'},
+            }
+        })
+        self.assertEqual(child.__schema__, {
+            'allOf': [
+                {'$ref': '#/definitions/Parent'},
+                {
+                    'properties': {
+                        'extra': {'type': 'string'}
+                    }
+                }
+            ]
         })
 
-        self.assertIn('Person', self.api.models)
-        self.assertIn('Child', self.api.models)
+    def test_inherit_from_class_from_multiple_parents(self):
+        grand_parent = Model('GrandParent', {
+            'grand_parent': fields.String,
+        })
+
+        parent = Model('Parent', {
+            'name': fields.String,
+            'age': fields.Integer,
+        })
+
+        child = Model.inherit('Child', grand_parent, parent, {
+            'extra': fields.String,
+        })
+
+        self.assertEqual(child.__schema__, {
+            'allOf': [
+                {'$ref': '#/definitions/GrandParent'},
+                {'$ref': '#/definitions/Parent'},
+                {
+                    'properties': {
+                        'extra': {'type': 'string'}
+                    }
+                }
+            ]
+        })
+
+    def test_inherit_from_instance_from_multiple_parents(self):
+        grand_parent = Model('GrandParent', {
+            'grand_parent': fields.String,
+        })
+
+        parent = Model('Parent', {
+            'name': fields.String,
+            'age': fields.Integer,
+        })
+
+        child = grand_parent.inherit('Child', parent, {
+            'extra': fields.String,
+        })
+
+        self.assertEqual(child.__schema__, {
+            'allOf': [
+                {'$ref': '#/definitions/GrandParent'},
+                {'$ref': '#/definitions/Parent'},
+                {
+                    'properties': {
+                        'extra': {'type': 'string'}
+                    }
+                }
+            ]
+        })
+
+    # def test_inherit_inline(self):
+    #     parent = Model('Person', {
+    #         'name': fields.String,
+    #         'age': fields.Integer,
+    #     })
+    #
+    #     child = self.api.inherit('Child', parent, {
+    #         'extra': fields.String,
+    #     })
+    #
+    #     Model('Output', {
+    #         'child': fields.Nested(child),
+    #         'children': fields.List(fields.Nested(child))
+    #     })
+    #
+    #     self.assertIn('Person', Models)
+    #     self.assertIn('Child', Models)
 
     def test_polymorph_inherit_common_ancestor(self):
         class Child1:
@@ -360,16 +487,16 @@ class ModelTestCase(TestCase):
         class Child2:
             pass
 
-        parent = self.api.model('Person', {
+        parent = Model('Person', {
             'name': fields.String,
             'age': fields.Integer,
         })
 
-        child1 = self.api.inherit('Child1', parent, {
+        child1 = parent.inherit('Child1', {
             'extra1': fields.String,
         })
 
-        child2 = self.api.inherit('Child2', parent, {
+        child2 = parent.inherit('Child2', {
             'extra2': fields.String,
         })
 
@@ -378,7 +505,7 @@ class ModelTestCase(TestCase):
             Child2: child2,
         }
 
-        output = self.api.model('Output', {
+        output = Model('Output', {
             'child': fields.Polymorph(mapping)
         })
 
@@ -386,5 +513,37 @@ class ModelTestCase(TestCase):
         self.assertEqual(output.__schema__, {
             'properties': {
                 'child': {'$ref': '#/definitions/Person'},
+            }
+        })
+
+
+class ModelDeprecattionsTest(TestCase):
+    def test_extend_is_deprecated(self):
+        parent = Model('Parent', {
+            'name': fields.String,
+            'age': fields.Integer,
+            'birthdate': fields.DateTime,
+        })
+
+        with self.assert_warning(DeprecationWarning):
+            child = parent.extend('Child', {
+                'extra': fields.String,
+            })
+
+        self.assertEqual(child.__schema__, {
+            'properties': {
+                'name': {
+                    'type': 'string'
+                },
+                'age': {
+                    'type': 'integer'
+                },
+                'birthdate': {
+                    'type': 'string',
+                    'format': 'date-time'
+                },
+                'extra': {
+                    'type': 'string'
+                }
             }
         })